+++ utils/__init__.py
@@ -1,4 +1,6 @@
-from .utils import *
+from .classes import *
+from .system import *
+from .scenarios import *
 
 from utils.lang.en import *
 from utils.lang.ru import *